run_experiments.py: do not auto-create kernel .config.tmpl
The template was previously copied from current .config if it did not exist.
But if the user did not use the template, this means that we might
unexpectedly pick up a previously copied file as baseline for a new
campaign, ignoring the user's customization of .config.

Also add numbered backups for the automated backup copy of .config.